Equipo 2/Fase 2/Evidencias Proyecto/Evidencias de Sistema/Base de datos/backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.7.2
asttokens==3.0.0
astunparse==1.6.3
bcrypt==3.2.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.2
cryptography==46.0.3
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
filelock==3.13.1
flatbuffers==25.9.23
gast==0.6.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.15.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
Pillow==10.1.0
pip-autoremove==0.10.0
pipdeptree==2.29.0
platformdirs==4.0.0
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2023.3.post1
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.48.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tornado==6.5.1
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.24.7
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.2
